+++ assembly-syntax/src/ast/procedure/name.rs
@@ -70,6 +70,22 @@ impl FromStr for QualifiedProcedureName {
     }
 }
 
+impl TryFrom<&str> for QualifiedProcedureName {
+    type Error = Report;
+
+    fn try_from(name: &str) -> Result<Self, Self::Error> {
+        Self::from_str(name)
+    }
+}
+
+impl TryFrom<String> for QualifiedProcedureName {
+    type Error = Report;
+
+    fn try_from(name: String) -> Result<Self, Self::Error> {
+        Self::from_str(&name)
+    }
+}
+
 impl Eq for QualifiedProcedureName {}
 
 impl PartialEq for QualifiedProcedureName {